[Addresses #59094046]MemberCall: list should only take params
Why: because passing the admin_user_id separately and then merging it into
the hash that's sent to the API is repetitive.  The admin_user_id is
already in the hash most of the time (example: it was already in the hash
for all of the spec tests written for it).